feat:画像変更、コンポーネント区分、ドキュメントの追加
