+++ Challenge_Panel/public/DOCUMENTATION.md
@@ -1,0 +1,282 @@
+# Challenge Panel - ドキュメント
+
+## 概要
+
+**Challenge Panel** は「部室王」チャレンジシステムのフロントエンドアプリケーションです。  
+部室への来室回数に応じてタスクが自動的にアンロックされ、3×3のアートパネルが徐々に公開されていくゲーム風UIを提供します。
+
+---
+
+## 技術スタック
+
+| 技術       | バージョン | 用途                         |
+| ---------- | ---------- | ---------------------------- |
+| Vue.js     | 3.5.0      | フロントエンドフレームワーク |
+| TypeScript | 5.8.0      | 型安全な開発                 |
+| Vite       | 7.3.0      | 開発サーバー・ビルドツール   |
+| Pinia      | 3.0.3      | 状態管理                     |
+| Vue Router | 4.5.1      | ルーティング                 |
+
+---
+
+## ディレクトリ構成
+
+```
+Challenge_Panel/
+├── public/                    # 静的ファイル
+│   └── DOCUMENTATION.md       # このドキュメント
+├── src/
+│   ├── assets/                # 画像・静的リソース
+│   │   └── 1.png ~ 9.png      # アートパネル用画像（9枚）
+│   ├── components/            # 再利用可能なコンポーネント
+│   │   └── ImageGrid.vue      # 3×3画像グリッドコンポーネント
+│   ├── stores/                # Pinia ストア
+│   │   └── taskStore.ts       # タスク管理・API連携
+│   ├── views/                 # ページコンポーネント
+│   │   ├── HomeView.vue       # ヘッダー（メインレイアウト）
+│   │   └── MainView.vue       # メインコンテンツ
+│   ├── router/
+│   │   └── index.ts           # ルーター設定
+│   ├── App.vue                # ルートコンポーネント
+│   └── main.ts                # エントリーポイント
+├── vite.config.ts             # Vite設定（APIプロキシ含む）
+├── tsconfig.json              # TypeScript設定
+└── package.json               # 依存関係
+```
+
+---
+
+## 各ファイルの役割
+
+### `/src/views/HomeView.vue`
+
+**ヘッダーコンポーネント**
+
+- アプリケーション全体のレイアウトを管理
+- ヘッダー部分のみを担当
+  - タイトル「部★室★王」
+  - 現在の部室王表示
+  - リアルタイム時計（1秒ごとに更新）
+- `MainView` コンポーネントを呼び出し
+
+### `/src/views/MainView.vue`
+
+**メインコンテンツコンポーネント**
+
+- 3つの主要セクションを表示:
+  1. **ART PANEL** - 3×3画像グリッド（中央）
+  2. **NEXT TASK** - 次にクリアすべきタスク（右上）
+  3. **STATUS** - 現在のカウント・進捗率（右下）
+- タスクストアと連携してデータを表示
+- 全タスク完了時は「ALL CLEAR!」アニメーションを表示
+
+### `/src/components/ImageGrid.vue`
+
+**画像グリッドコンポーネント**
+
+- Canvas APIを使用して3×3グリッドを描画
+- 9枚の画像（1.png～9.png）を表示
+- 未完了タスクの画像にはオーバーレイを適用
+- 10秒ごとにAPIからデータを取得して更新
+
+**主な機能:**
+
+- `drawGrid()`: グリッド全体を描画
+- `loadImage()`: 画像の非同期読み込み
+- 完了したタスクのみ画像が鮮明に表示される
+
+### `/src/stores/taskStore.ts`
+
+**タスク管理ストア（Pinia）**
+
+- 9つのタスクを管理
+- 各タスクには以下の情報が含まれる:
+  - `id`: タスクID（0-8）
+  - `name`: タスク名
+  - `unlockCount`: アンロックに必要な来室人数
+  - `completed`: 完了状態
+  - `imageFile`: 対応する画像ファイル名
+
+**主な機能:**
+
+- `fetchCount()`: APIから現在の来室人数を取得
+- `updateTasksByCount()`: 来室人数に基づいてタスクの完了状態を更新
+
+**Computed プロパティ:**
+
+- `completedCount`: 完了タスク数
+- `progressPercentage`: 進捗率（%）
+
+### `/vite.config.ts`
+
+**Vite設定ファイル**
+
+- APIプロキシ設定（CORS対策）
+- `/api/*` へのリクエストを `https://hacku2025.vercel.app` に転送
+
+---
+
+## API連携
+
+### エンドポイント
+
+```
+GET /api/attendance/today-count
+```
+
+### リクエストヘッダー
+
+```
+HARDWARE_API_KEY: Procon2025
+```
+
+### レスポンス
+
+```json
+{
+  "count": 5
+}
+```
+
+### プロキシ設定
+
+開発環境ではCORS問題を回避するため、Viteプロキシを使用:
+
+```typescript
+// vite.config.ts
+server: {
+  proxy: {
+    '/api': {
+      target: 'https://hacku2025.vercel.app',
+      changeOrigin: true,
+    }
+  }
+}
+```
+
+---
+
+## タスク一覧
+
+| タスクID | タスク名     | 必要人数 | 画像ファイル |
+| -------- | ------------ | -------- | ------------ |
+| 1        | 最初の来訪者 | 1人      | 1.png        |
+| 2        | 常連への道   | 2人      | 2.png        |
+| 3        | 部室の住人   | 3人      | 3.png        |
+| 4        | 中堅メンバー | 4人      | 4.png        |
+| 5        | ベテラン部員 | 5人      | 5.png        |
+| 6        | 部室マスター | 6人      | 6.png        |
+| 7        | 伝説の始まり | 7人      | 7.png        |
+| 8        | 部室の守護者 | 8人      | 8.png        |
+| 9        | 部室王       | 9人      | 9.png        |
+
+---
+
+## 動作フロー
+
+```
+┌─────────────────────────────────────────────────────┐
+│                    アプリ起動                        │
+└─────────────────────────────────────────────────────┘
+                         │
+                         ▼
+┌─────────────────────────────────────────────────────┐
+│              API から来室人数を取得                   │
+│         GET /api/attendance/today-count             │
+└─────────────────────────────────────────────────────┘
+                         │
+                         ▼
+┌─────────────────────────────────────────────────────┐
+│           タスクの完了状態を更新                      │
+│     count >= task.unlockCount → completed = true    │
+└─────────────────────────────────────────────────────┘
+                         │
+                         ▼
+┌─────────────────────────────────────────────────────┐
+│              UIを再描画                              │
+│  - 画像グリッド: 完了タスクの画像を表示              │
+│  - NEXT TASK: 次の未完了タスクを表示                 │
+│  - STATUS: カウント・進捗率を更新                    │
+└─────────────────────────────────────────────────────┘
+                         │
+                         ▼
+            ┌────────────────────────┐
+            │   10秒後に再度取得     │
+            │      (ループ)          │
+            └────────────────────────┘
+```
+
+---
+
+## 開発・実行
+
+### 開発サーバー起動
+
+```bash
+cd Challenge_Panel
+npm install
+npm run dev
+```
+
+### ビルド
+
+```bash
+npm run build
+```
+
+### プレビュー
+
+```bash
+npm run preview
+```
+
+---
+
+## 画像アセット
+
+`/src/assets/` ディレクトリに以下の9枚の画像を配置してください:
+
+- `1.png` - タスク1用画像
+- `2.png` - タスク2用画像
+- `3.png` - タスク3用画像
+- `4.png` - タスク4用画像
+- `5.png` - タスク5用画像
+- `6.png` - タスク6用画像
+- `7.png` - タスク7用画像
+- `8.png` - タスク8用画像
+- `9.png` - タスク9用画像
+
+画像は3×3グリッドとして表示されるため、正方形またはアスペクト比1:1の画像を推奨します。
+
+---
+
+## UI デザイン
+
+- **テーマ**: ゲーム風（スマブラ風）のカクカクしたデザイン
+- **カラースキーム**:
+  - メイン: `#00cc66` (緑)
+  - 背景: `#0a0a0a` (黒)
+  - アクセント: `#ffd700` (金)
+- **特徴**:
+  - 角張った形状（`clip-path` 使用）
+  - グロー効果（`box-shadow`, `text-shadow`）
+  - アニメーション（ALL CLEAR時のバウンス効果）
+
+---
+
+## 注意事項
+
+1. **画像ファイル**: 実行前に必ず1.png～9.pngをassetsフォルダに配置してください
+2. **API**: 開発環境ではViteプロキシを使用します。本番環境では適切なCORS設定が必要です
+3. **ポーリング**: 10秒間隔でAPIを呼び出します。本番環境では間隔の調整を検討してください
+
+---
+
+## 更新履歴
+
+- **2025/12/18**: 初期バージョン作成
+  - Vue 3 + TypeScript + Vite環境構築
+  - 3×3画像グリッド実装
+  - API連携・自動アンロック機能
+  - ゲーム風UI実装
+  - HomeView/MainView分離
